Refactor e2e tests file structure
