+++ test/catalog/attribute-types.e2e-spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { AppModule } from '../src/app.module';
-import { Role } from '../src/users/entities/role.enum';
-import { TestUsersService } from './utils/test-users/test-users.service';
-import { TestUsersModule } from './utils/test-users/test-users.module';
-import { DtoGeneratorService } from './utils/dto-generator/dto-generator.service';
-import { AttributeType } from '../src/catalog/entities/attribute-type.entity';
-import { AttributeTypeDto } from '../src/catalog/dto/attribute-type.dto';
-import { setupRbacTests } from './utils/setup-rbac-tests';
+import { AppModule } from '../../src/app.module';
+import { Role } from '../../src/users/entities/role.enum';
+import { TestUsersService } from '../utils/test-users/test-users.service';
+import { TestUsersModule } from '../utils/test-users/test-users.module';
+import { DtoGeneratorService } from '../utils/dto-generator/dto-generator.service';
+import { AttributeType } from '../../src/catalog/entities/attribute-type.entity';
+import { AttributeTypeDto } from '../../src/catalog/dto/attribute-type.dto';
+import { setupRbacTests } from '../utils/setup-rbac-tests';
 
 describe('AttributeTypesController (e2e)', () => {
   let app: INestApplication;
