+++ test/sales/orders.e2e-spec.ts
@@ -1,24 +1,24 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { AppModule } from '../src/app.module';
-import { Role } from '../src/users/entities/role.enum';
-import { TestUsersService } from './utils/test-users/test-users.service';
-import { TestUsersModule } from './utils/test-users/test-users.module';
-import { DtoGeneratorService } from './utils/dto-generator/dto-generator.service';
-import { ProductCreateDto } from '../src/catalog/dto/product-create.dto';
-import { Product } from '../src/catalog/entities/product.entity';
-import { setupRbacTests } from './utils/setup-rbac-tests';
-import { OrderCreateDto } from '../src/sales/dto/order-create.dto';
-import { Order } from '../src/sales/entities/order.entity';
-import { OrderUpdateDto } from '../src/sales/dto/order-update.dto';
-import { DeliveryMethodDto } from '../src/sales/dto/delivery-method.dto';
-import { DeliveryMethod } from '../src/sales/entities/delivery-method.entity';
-import { OrderDeliveryDto } from '../src/sales/dto/order-delivery.dto';
-import { PaymentMethod } from '../src/sales/entities/payment-method.entity';
-import { PaymentMethodDto } from '../src/sales/dto/payment-method.dto';
-import { OrderPaymentDto } from '../src/sales/dto/order-payment.dto';
-import { OrderStatus } from '../src/sales/entities/order-status.enum';
+import { AppModule } from '../../src/app.module';
+import { Role } from '../../src/users/entities/role.enum';
+import { TestUsersService } from '../utils/test-users/test-users.service';
+import { TestUsersModule } from '../utils/test-users/test-users.module';
+import { DtoGeneratorService } from '../utils/dto-generator/dto-generator.service';
+import { ProductCreateDto } from '../../src/catalog/dto/product-create.dto';
+import { Product } from '../../src/catalog/entities/product.entity';
+import { setupRbacTests } from '../utils/setup-rbac-tests';
+import { OrderCreateDto } from '../../src/sales/dto/order-create.dto';
+import { Order } from '../../src/sales/entities/order.entity';
+import { OrderUpdateDto } from '../../src/sales/dto/order-update.dto';
+import { DeliveryMethodDto } from '../../src/sales/dto/delivery-method.dto';
+import { DeliveryMethod } from '../../src/sales/entities/delivery-method.entity';
+import { OrderDeliveryDto } from '../../src/sales/dto/order-delivery.dto';
+import { PaymentMethod } from '../../src/sales/entities/payment-method.entity';
+import { PaymentMethodDto } from '../../src/sales/dto/payment-method.dto';
+import { OrderPaymentDto } from '../../src/sales/dto/order-payment.dto';
+import { OrderStatus } from '../../src/sales/entities/order-status.enum';
 
 describe.only('OrdersController (e2e)', () => {
   let app: INestApplication;
